Getting rid of IVMapManager
